+++ cvxpy/tests/test_problem.py
@@ -1275,4 +1275,3 @@ class TestProblem(BaseTest):
         prob.solve()
         x = x.value
         self.assertTrue(__builtins__['abs'](1.7*x**.7 - 2.3*x**-3.3 - .45*x**-.55) <= 1e-3)
-