+++ CHANGELOG.md
@@ -19,7 +19,7 @@ Version 1.0 (targets)
 
 Version 0.2.21 (next release)
 -----------------------------
-* Made operator overloading work with scipy sparse matrices.
+* Made operator overloading work with scipy sparse matrices (with scipy 0.15).
 * Removed Expression shape function.
 * Removed Expression __array__ function.